Containerized-FullStack-Webapp/requirements.txt
dnspython  
flask
mysql-connector-python
flask_wtf
wtforms
flask_mail
